requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
Click==7.0
decorator==4.4.1
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
importlib-metadata==0.23
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.18.2
more-itertools==7.2.0
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.2.2
python-dotenv==0.10.3
six==1.12.0
traitlets==4.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
